Don't send (unused) APP_CMD_KEY/TOUCH_EVENTs
We already have an `InputAvailable` wakeup event to ensure the
application will poll input within a finite amount of time after new
input arrives.

The advantage of `InputAvailable` is that it's more consistent with
`NativeActivity` and it doesn't require us to write an event for _every_
input event if we know the application has already been notified (and
has not yet polled for new input).